fix: move some unit tests from Checkout.test to ContactInformation.test
