+++ src/components/ChangePlan/Checkout/Checkout.test.js
@@ -21,7 +21,6 @@ describe('Checkout component', () => {
     phone: '+542494228090',
     company: 'Test',
     industry: 'IT',
-    completed: true,
   };
 
   const dependencies = {
@@ -61,452 +60,18 @@ describe('Checkout component', () => {
     expect(getAllByText('checkoutProcessForm.contact_information_title').length).toBe(2);
   });
 
-  it('should show messages for empty required fields', async () => {
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getAllByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: '' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: '' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: '' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: '' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: '' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: '' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: '' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getAllByText('validation_messages.error_required_field').length).toBe(8);
-  });
-
-  it('should show error message if firstname field is empty', async () => {
+  it('should show contact information, billing information and payment method header', async () => {
     // Arrange
     let result;
     await act(async () => {
       result = render(<CheckoutElement />);
     });
 
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: '' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: '1234' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'buenos aires' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
+    const { getAllByText } = result;
 
     // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if lastname field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: '' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: '1234' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'buenos aires' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if address field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: '' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'buenos aires' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if city field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: '' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'buenos aires' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if province field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: '' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if country field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'province 1' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: '' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if phone field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'province 1' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show error message if industry field is empty', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'Test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'Test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'tandil' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'pronvince 1' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: '' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '+54 223 655-8877' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('validation_messages.error_required_field')).toBeInTheDocument();
-  });
-
-  it('should show success message if the submit was succesfully', async () => {
-    // Arrange
-    let result;
-    await act(async () => {
-      result = render(<CheckoutElement />);
-    });
-
-    const { container, getByText } = result;
-
-    // Act
-    act(() => {
-      const inputFirstname = container.querySelector('input#firstname');
-      fireEvent.change(inputFirstname, { target: { value: 'test' } });
-
-      const inputLastname = container.querySelector('input#lastname');
-      fireEvent.change(inputLastname, { target: { value: 'test' } });
-
-      const inputAddress = container.querySelector('input#address');
-      fireEvent.change(inputAddress, { target: { value: 'address 1' } });
-
-      const inputCity = container.querySelector('input#city');
-      fireEvent.change(inputCity, { target: { value: 'city' } });
-
-      const inputProvince = container.querySelector('input#province');
-      fireEvent.change(inputProvince, { target: { value: 'province' } });
-
-      const selectCountry = container.querySelector('select#country');
-      fireEvent.change(selectCountry, { target: { value: 'ar' } });
-
-      const selectIndustry = container.querySelector('select#industry');
-      fireEvent.change(selectIndustry, { target: { value: 'dplr1' } });
-
-      const inputPhone = container.querySelector('input#phone');
-      fireEvent.change(inputPhone, { target: { value: '2494222222' } });
-    });
-
-    await act(async () => {
-      const submitButton = container.querySelector('button[type="submit"]');
-      fireEvent.submit(submitButton);
-    });
-
-    // Assert
-    expect(getByText('checkoutProcessForm.success_msg')).toBeInTheDocument();
+    expect(getAllByText('checkoutProcessForm.contact_information_title').length).toBe(2);
+    expect(getAllByText('checkoutProcessForm.billing_information_title').length).toBe(1);
+    expect(getAllByText('checkoutProcessForm.payment_method_title').length).toBe(1);
   });
 });
